enable_star=1 by default enable_star marked as deprecated fixed enable_star=1 + wordforms queries: now we do use wordforms is such cases not dict=keywords is more default than before code cleanup
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4114 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxsearch.cpp
@@ -3646,7 +3646,7 @@ DWORD GetQposMask ( const CSphVector<ExtNode_i *> & dQwords )
 	for ( int i=1; i<dQwords.GetLength(); i++ )
 	{
 		int iQposDelta = dQwords[i]->m_iAtomPos - iQposBase;
-		assert ( iQposDelta<sizeof(uQposMask)*8 );
+		assert ( iQposDelta<(int)sizeof(uQposMask)*8 );
 		uQposMask |= ( 1 << iQposDelta );
 	}
 
@@ -4020,7 +4020,7 @@ inline bool FSMmultinear::HitFSM ( const ExtHit_t* pHit, ExtHit_t* dTarget )
 				ARRAY_FOREACH ( i, m_dNpos )
 				{
 					int iQposDelta = m_dNpos[i] - dTarget->m_uQuerypos;
-					assert ( iQposDelta<sizeof(dTarget->m_uQposMask)*8 );
+					assert ( iQposDelta<(int)sizeof(dTarget->m_uQposMask)*8 );
 					dTarget->m_uQposMask |= ( 1 << iQposDelta );
 				}
 			}
